Break backwards compatibility to overcome TLS vulnerabilities, but allow clients to opt-out
The CONC-648 and CONC-654 vulnerabilities cannot be addressed
on the client side in a way that is fully backwards-compatible with
existing servers:

1. CONC-648:

   Per Sergei Golubchik's research
   (https://jira.mariadb.org/browse/CONC-648?focusedCommentId=263341#comment-263341),
   there are numerous legitimate errors which servers sometimes send
   prior to the TLS handshake.

2. CONC-654:

   There is a server bug which must be fixed
   (https://jira.mariadb.org/browse/MDEV-31585) in order for the
   client and server to be able to establish a correctly-configured
   session with the client-side vulnerability fixed.

At the same time, if we allow the client to continue silently
downgrading itself (= "exhibiting these vulnerabilities") in the name
of backwards-compatibility, end users will have no motivation to
upgrade.

Instead, we should change the client to avoid these vulnerabilities
unconditionally.

At the same time, add a special client option
`backwards-compatible-insecure-ssl` (alias
`backwards-compatible-insecure-tls`) to allow the client to continue
implementing the old, insecure behavior...  but with noisy warnings every
time this behavior is required.

All new code of the whole pull request, including one or several files
that are either new files or modified ones, are contributed under the
BSD-new license. I am contributing on behalf of my employer Amazon Web
Services, Inc.